# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RoomCreateRequest": {
        "properties": {
          "room_type": {
            "default": "social",
            "maxLength": 50,
            "minLength": 1,
            "title": "Room Type",
            "type": "string"
          }
        },
        "title": "RoomCreateRequest",
        "type": "object"
      },
      "RoomCreateResponse": {
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "created_by_user_id": {
            "title": "Created By User Id",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "room_type": {
            "title": "Room Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "room_type",
          "created_by_user_id",
          "created_at"
        ],
        "title": "RoomCreateResponse",
        "type": "object"
      },
      "RoomInviteRequest": {
        "properties": {
          "target_user_id": {
            "minLength": 1,
            "title": "Target User Id",
            "type": "string"
          }
        },
        "required": [
          "target_user_id"
        ],
        "title": "RoomInviteRequest",
        "type": "object"
      },
      "RoomInviteResponse": {
        "properties": {
          "actor_user_id": {
            "title": "Actor User Id",
            "type": "string"
          },
          "room_id": {
            "title": "Room Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "target_user_id": {
            "title": "Target User Id",
            "type": "string"
          }
        },
        "required": [
          "room_id",
          "actor_user_id",
          "target_user_id",
          "status"
        ],
        "title": "RoomInviteResponse",
        "type": "object"
      },
      "RoomJoinResponse": {
        "properties": {
          "actor_user_id": {
            "title": "Actor User Id",
            "type": "string"
          },
          "room_id": {
            "title": "Room Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "target_user_id": {
            "title": "Target User Id",
            "type": "string"
          }
        },
        "required": [
          "room_id",
          "actor_user_id",
          "target_user_id",
          "status"
        ],
        "title": "RoomJoinResponse",
        "type": "object"
      },
...
      "ScreenshotRequest": {
        "properties": {
          "emit_ws_event": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "description": "Override privacy_mode default WS behavior",
            "title": "Emit Ws Event"
          },
          "image_base64": {
            "minLength": 1,
            "title": "Image Base64",
            "type": "string"
          },
          "privacy_mode": {
            "default": "strict",
            "description": "strict: only return suggestion; standard: also append WS SUGGESTION event (no screenshot content)",
            "enum": [
              "strict",
              "standard"
            ],
            "title": "Privacy Mode",
            "type": "string"
          },
          "save_id": {
            "minLength": 1,
            "title": "Save Id",
            "type": "string"
          }
        },
        "required": [
          "save_id",
          "image_base64"
        ],
        "title": "ScreenshotRequest",
        "type": "object"
      },
      "ScreenshotResponse": {
        "properties": {
          "suggestion": {
            "title": "Suggestion",
            "type": "string"
          }
        },
        "required": [
          "suggestion"
        ],
        "title": "ScreenshotResponse",
        "type": "object"
      },
      "SensorEventRequest": {
        "properties": {
          "app_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "App Name"
          },
          "event_type": {
            "enum": [
              "clipboard",
              "idle"
            ],
            "title": "Event Type",
            "type": "string"
          },
          "idle_seconds": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idle Seconds"
          },
          "save_id": {
            "minLength": 1,
            "title": "Save Id",
            "type": "string"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          }
        },
        "required": [
          "save_id",
          "event_type"
        ],
        "title": "SensorEventRequest",
        "type": "object"
      },
      "SensorEventResponse": {
        "properties": {
          "category": {
            "enum": [
              "translation",
              "care",
              "generic"
            ],
            "title": "Category",
            "type": "string"
          },
          "suggestion": {
            "minLength": 1,
            "title": "Suggestion",
            "type": "string"
          }
        },
        "required": [
          "suggestion",
          "category"
        ],
        "title": "SensorEventResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "para-server",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/sensors/event": {
      "post": {
        "operationId": "sensors_event",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SensorEventRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SensorEventResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Sensors Event",
        "tags": [
          "sensors"
        ]
      }
    },
    "/api/v1/sensors/screenshot": {
      "post": {
        "operationId": "sensors_screenshot",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScreenshotRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScreenshotResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Sensors Screenshot",
        "tags": [
          "sensors"
        ]
      }
    },
    "/api/v1/social/rooms": {
      "post": {
        "operationId": "social_rooms_create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoomCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomCreateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Social Rooms Create",
        "tags": [
          "social"
        ]
      }
    },
    "/api/v1/social/rooms/{room_id}/invite": {
      "post": {
        "operationId": "social_rooms_invite",
        "parameters": [
          {
            "in": "path",
            "name": "room_id",
            "required": true,
            "schema": {
              "title": "Room Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoomInviteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomInviteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Social Rooms Invite",
        "tags": [
          "social"
        ]
      }
    },
    "/api/v1/social/rooms/{room_id}/join": {
      "post": {
        "operationId": "social_rooms_join",
        "parameters": [
          {
            "in": "path",
            "name": "room_id",
            "required": true,
            "schema": {
              "title": "Room Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomJoinResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Social Rooms Join",
        "tags": [
          "social"
        ]
      }
    },
...
  }
}